Upgrade to AWS SDK v4 (#2816)
* upgrade to v4 and see what breaks

* Collections are by default null

* Fix dispatcher

* Fix InputQueue pump

* Fix delayed message pump

* Fix test builder

* Simplify policy extensions

* Fix policy scrubber

* Fix sqs prepared message

* Cosmetics on prepared message

* Cosmetics on subscription manager tests

* Approval files

* Sns prepared message alignment

* Delayed message pump check for null received messages

* Temporary fix for clock skew

* Fix mock sqs client

* Remove no longer needed methods on fake clients

* Null checks in command line

* Bump sdk to bug fixed version

* Fix failing test

* Fix lifecycle policy rules expression

* Fix bucket creation when creating queues

* Prevent possible NRE

* Fix message ID and transport message extraction checking for null collections

* Fix message attribute access

* Static

* DetermineServiceOperationEndpoint doesn't require an already executed request

* Check the endpointUrl is not null

---------

+++ src/NServiceBus.Transport.SQS/Receiving/DelayedMessagesPump.cs
@@ -6,6 +6,7 @@ using System.Linq;
 using System.Text;
 using System.Threading;
 using System.Threading.Tasks;
+using Amazon.Runtime;
 using Amazon.SQS;
 using Amazon.SQS.Model;
 using Extensions;
@@ -87,6 +88,8 @@ class DelayedMessagesPump(string receiveAddress, IAmazonSQS sqsClient, QueueCach
             MessageAttributeNames = ["All"]
         };
 
+        endpointUrl = sqsClient.DetermineServiceOperationEndpoint(receiveDelayedMessagesRequest).URL;
+
         // Task.Run() so the call returns immediately instead of waiting for the first await or return down the call stack
         pumpTask = Task.Run(() => ConsumeDelayedMessagesAndSwallowExceptions(receiveDelayedMessagesRequest, tokenSource.Token), CancellationToken.None);
     }
@@ -134,17 +137,16 @@ class DelayedMessagesPump(string receiveAddress, IAmazonSQS sqsClient, QueueCach
     internal async Task ConsumeDelayedMessages(ReceiveMessageRequest request, CancellationToken cancellationToken = default)
     {
         var receivedMessages = await sqsClient.ReceiveMessageAsync(request, cancellationToken).ConfigureAwait(false);
-        if (receivedMessages.Messages.Count == 0)
+        if (receivedMessages.Messages is { Count: > 0 })
         {
-            return;
+            // In some unit test the pump is not started, with the consequence that the endpoint URL is never evaluated
+            var clockCorrection = endpointUrl == null ? TimeSpan.Zero : CorrectClockSkew.GetClockCorrectionForEndpoint(endpointUrl);
+            var preparedMessages = PrepareMessages(receivedMessages, clockCorrection, cancellationToken);
+
+            cancellationToken.ThrowIfCancellationRequested();
+
+            await BatchDispatchPreparedMessages(preparedMessages, cancellationToken).ConfigureAwait(false);
         }
-
-        var clockCorrection = sqsClient.Config.ClockOffset;
-        var preparedMessages = PrepareMessages(receivedMessages, clockCorrection, cancellationToken);
-
-        cancellationToken.ThrowIfCancellationRequested();
-
-        await BatchDispatchPreparedMessages(preparedMessages, cancellationToken).ConfigureAwait(false);
     }
 
     IReadOnlyCollection<SqsReceivedDelayedMessage> PrepareMessages(ReceiveMessageResponse receivedMessages, TimeSpan clockCorrection, CancellationToken cancellationToken)
@@ -157,13 +159,13 @@ class DelayedMessagesPump(string receiveAddress, IAmazonSQS sqsClient, QueueCach
             preparedMessages ??= new List<SqsReceivedDelayedMessage>(receivedMessages.Messages.Count);
             long delaySeconds = 0;
 
-            if (receivedMessage.MessageAttributes.TryGetValue(TransportHeaders.DelaySeconds, out var delayAttribute))
+            if (receivedMessage.MessageAttributes?.TryGetValue(TransportHeaders.DelaySeconds, out var delayAttribute) is true)
             {
                 long.TryParse(delayAttribute.StringValue, out delaySeconds);
             }
 
-            string originalMessageId = null;
-            if (receivedMessage.MessageAttributes.TryGetValue(Headers.MessageId, out var messageIdAttribute))
+            string originalMessageId = receivedMessage.MessageId;
+            if (receivedMessage.MessageAttributes?.TryGetValue(Headers.MessageId, out var messageIdAttribute) is true)
             {
                 originalMessageId = messageIdAttribute.StringValue;
             }
@@ -292,39 +294,45 @@ class DelayedMessagesPump(string receiveAddress, IAmazonSQS sqsClient, QueueCach
         await deletionTask.ConfigureAwait(false);
     }
 
-    async Task ChangeVisibilityOfDelayedMessagesThatFailedBatchDeliveryInBatches(SqsBatchEntry batch, SendMessageBatchResponse result, int batchNumber, int totalBatches, CancellationToken cancellationToken)
+    async Task ChangeVisibilityOfDelayedMessagesThatFailedBatchDeliveryInBatches(SqsBatchEntry batch,
+        SendMessageBatchResponse result, int batchNumber, int totalBatches, CancellationToken cancellationToken)
     {
-        try
+        if (result.Failed is { Count: > 0 })
         {
-            List<ChangeMessageVisibilityBatchRequestEntry> changeVisibilityBatchRequestEntries = null;
-            foreach (var failed in result.Failed)
+            try
             {
-                changeVisibilityBatchRequestEntries ??= new List<ChangeMessageVisibilityBatchRequestEntry>(result.Failed.Count);
-                var preparedMessage = (SqsReceivedDelayedMessage)batch.PreparedMessagesBydId[failed.Id];
-                // need to reuse the previous batch entry ID so that we can map again in failure scenarios, this is fine given that IDs only need to be unique per request
-                changeVisibilityBatchRequestEntries.Add(new ChangeMessageVisibilityBatchRequestEntry(failed.Id, preparedMessage.ReceiptHandle)
+                var changeVisibilityBatchRequestEntries = new List<ChangeMessageVisibilityBatchRequestEntry>(result.Failed.Count);
+                foreach (var failed in result.Failed)
                 {
-                    VisibilityTimeout = 0
-                });
-            }
 
-            if (changeVisibilityBatchRequestEntries != null)
-            {
+                    var preparedMessage = (SqsReceivedDelayedMessage)batch.PreparedMessagesBydId[failed.Id];
+                    // need to reuse the previous batch entry ID so that we can map again in failure scenarios, this is fine given that IDs only need to be unique per request
+                    changeVisibilityBatchRequestEntries.Add(
+                        new ChangeMessageVisibilityBatchRequestEntry(failed.Id, preparedMessage.ReceiptHandle)
+                        {
+                            VisibilityTimeout = 0
+                        });
+                }
+
                 if (Logger.IsDebugEnabled)
                 {
                     var message = batch.PreparedMessagesBydId.Values.First();
 
-                    Logger.Debug($"Changing delayed message visibility for batch '{batchNumber}/{totalBatches}' with message ids '{string.Join(", ", batch.PreparedMessagesBydId.Values.Select(v => v.MessageId))}' for destination {message.Destination}");
+                    Logger.Debug(
+                        $"Changing delayed message visibility for batch '{batchNumber}/{totalBatches}' with message ids '{string.Join(", ", batch.PreparedMessagesBydId.Values.Select(v => v.MessageId))}' for destination {message.Destination}");
                 }
 
-                var changeVisibilityResult = await sqsClient.ChangeMessageVisibilityBatchAsync(new ChangeMessageVisibilityBatchRequest(delayedDeliveryQueueUrl, changeVisibilityBatchRequestEntries), cancellationToken)
+                var changeVisibilityResult = await sqsClient.ChangeMessageVisibilityBatchAsync(
+                        new ChangeMessageVisibilityBatchRequest(delayedDeliveryQueueUrl,
+                            changeVisibilityBatchRequestEntries), cancellationToken)
                     .ConfigureAwait(false);
 
                 if (Logger.IsDebugEnabled)
                 {
                     var message = batch.PreparedMessagesBydId.Values.First();
 
-                    Logger.Debug($"Changed delayed message visibility for batch '{batchNumber}/{totalBatches}' with message ids '{string.Join(", ", batch.PreparedMessagesBydId.Values.Select(v => v.MessageId))}' for destination {message.Destination}");
+                    Logger.Debug(
+                        $"Changed delayed message visibility for batch '{batchNumber}/{totalBatches}' with message ids '{string.Join(", ", batch.PreparedMessagesBydId.Values.Select(v => v.MessageId))}' for destination {message.Destination}");
                 }
 
                 if (Logger.IsDebugEnabled && changeVisibilityResult.Failed.Count > 0)
@@ -338,32 +346,32 @@ class DelayedMessagesPump(string receiveAddress, IAmazonSQS sqsClient, QueueCach
                     Logger.Debug($"Changing visibility failed for {builder}");
                 }
             }
-        }
-        catch (Exception ex) when (!ex.IsCausedBy(cancellationToken))
-        {
-            var builder = new StringBuilder();
-            foreach (var failed in result.Failed)
+            catch (Exception ex) when (!ex.IsCausedBy(cancellationToken))
             {
-                builder.AppendLine($"{failed.Id}: {failed.Message} | {failed.Code} | {failed.SenderFault}");
-            }
+                var builder = new StringBuilder();
+                foreach (var failed in result.Failed)
+                {
+                    builder.AppendLine($"{failed.Id}: {failed.Message} | {failed.Code} | {failed.SenderFault}");
+                }
 
-            Logger.Error($"Changing visibility failed for {builder}", ex);
+                Logger.Error($"Changing visibility failed for {builder}", ex);
+            }
         }
     }
 
     async Task DeleteDelayedMessagesThatWereDeliveredSuccessfullyAsBatchesInBatches(SqsBatchEntry batch, SendMessageBatchResponse result, int batchNumber, int totalBatches, CancellationToken cancellationToken)
     {
-        List<DeleteMessageBatchRequestEntry> deleteBatchRequestEntries = null;
-        foreach (var successful in result.Successful)
+        if (result.Successful is { Count: > 0 })
         {
-            deleteBatchRequestEntries ??= new List<DeleteMessageBatchRequestEntry>(result.Successful.Count);
-            var preparedMessage = (SqsReceivedDelayedMessage)batch.PreparedMessagesBydId[successful.Id];
-            // need to reuse the previous batch entry ID so that we can map again in failure scenarios, this is fine given that IDs only need to be unique per request
-            deleteBatchRequestEntries.Add(new DeleteMessageBatchRequestEntry(successful.Id, preparedMessage.ReceiptHandle));
-        }
+            var deleteBatchRequestEntries = new List<DeleteMessageBatchRequestEntry>(result.Successful.Count);
+            foreach (var successful in result.Successful)
+            {
 
-        if (deleteBatchRequestEntries != null)
-        {
+                var preparedMessage = (SqsReceivedDelayedMessage)batch.PreparedMessagesBydId[successful.Id];
+                // need to reuse the previous batch entry ID so that we can map again in failure scenarios, this is fine given that IDs only need to be unique per request
+                deleteBatchRequestEntries.Add(new DeleteMessageBatchRequestEntry(successful.Id, preparedMessage.ReceiptHandle));
+            }
+
             if (Logger.IsDebugEnabled)
             {
                 var message = batch.PreparedMessagesBydId.Values.First();
@@ -381,17 +389,16 @@ class DelayedMessagesPump(string receiveAddress, IAmazonSQS sqsClient, QueueCach
                 Logger.Debug($"Deleted delayed message for batch '{batchNumber}/{totalBatches}' with message ids '{string.Join(", ", batch.PreparedMessagesBydId.Values.Select(v => v.MessageId))}' for destination {message.Destination}");
             }
 
-            List<Task> deleteTasks = null;
-            foreach (var errorEntry in deleteResult.Failed)
+            if (deleteResult.Failed is { Count: > 0 })
             {
-                deleteTasks ??= new List<Task>(deleteResult.Failed.Count);
-                var messageToDeleteWithAnotherAttempt = (SqsReceivedDelayedMessage)batch.PreparedMessagesBydId[errorEntry.Id];
-                Logger.Info($"Retrying message deletion with MessageId {messageToDeleteWithAnotherAttempt.MessageId} that failed in batch '{batchNumber}/{totalBatches}' due to '{errorEntry.Message}'.");
-                deleteTasks.Add(DeleteMessage(messageToDeleteWithAnotherAttempt, cancellationToken));
-            }
+                var deleteTasks = new List<Task>(deleteResult.Failed.Count);
+                foreach (var errorEntry in deleteResult.Failed)
+                {
 
-            if (deleteTasks != null)
-            {
+                    var messageToDeleteWithAnotherAttempt = (SqsReceivedDelayedMessage)batch.PreparedMessagesBydId[errorEntry.Id];
+                    Logger.Info($"Retrying message deletion with MessageId {messageToDeleteWithAnotherAttempt.MessageId} that failed in batch '{batchNumber}/{totalBatches}' due to '{errorEntry.Message}'.");
+                    deleteTasks.Add(DeleteMessage(messageToDeleteWithAnotherAttempt, cancellationToken));
+                }
                 await Task.WhenAll(deleteTasks).ConfigureAwait(false);
             }
         }
@@ -417,6 +424,7 @@ class DelayedMessagesPump(string receiveAddress, IAmazonSQS sqsClient, QueueCach
     string delayedDeliveryQueueUrl;
     Task pumpTask;
     string inputQueueUrl;
+    string endpointUrl;
 
     // using the same logger for now
     static readonly ILog Logger = LogManager.GetLogger(typeof(MessagePump));
